Working Encoder

+++ firmware/libraries/NU_motor/QuadratureEncoder.h
@@ -7,7 +7,7 @@
 
 class Encoders{
   public:  
-    Encoders(byte pinA, byte pinB);
+    Encoders(uint32_t pinA, uint32_t pinB);
     static void interruptEncoder1(){
       if(Encoders::_instances[0] != NULL)
       Encoders::_instances[0]->encoderCount();
@@ -33,8 +33,8 @@ class Encoders{
     
   private:
     static uint8_t _whichEncoder;
-    uint8_t _encoderPINA;
-    uint8_t _encoderPINB;
+    uint32_t _encoderPINA;
+    uint32_t _encoderPINB;
     volatile long _encoderCount = 0;
     volatile int _lastEncoded = 0;
     volatile long _encoderErrors = 0;
